pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cicaddy-github"
version = "0.1.0"
description = "GitHub Actions plugin for cicaddy AI agent framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [{name = "Wayne Sun"}]

dependencies = [
    "cicaddy==0.2.0",
    "PyGithub>=2.1.0",
    "detect-secrets>=1.4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[project.entry-points."cicaddy.agents"]
github = "cicaddy_github.plugin:register_agents"

[project.entry-points."cicaddy.settings_loader"]
github = "cicaddy_github.config.settings:load_settings"

[project.entry-points."cicaddy.cli_args"]
github = "cicaddy_github.plugin:get_cli_args"

[project.entry-points."cicaddy.env_vars"]
github = "cicaddy_github.plugin:get_env_vars"

[project.entry-points."cicaddy.validators"]
github = "cicaddy_github.plugin:validate"

[tool.hatch.build.targets.wheel]
packages = ["src/cicaddy_github"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101", "B603", "B607"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
